psychometric/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
et-xmlfile==1.1.0
excel2json-3==0.1.6
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.1
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
xlrd==2.0.1
zipp==3.10.0
google-cloud-tasks==2.13.0
